backend/requirements.txt
# Flask Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7  # PostgreSQL adapter for production

# Redis and Celery for async tasks
redis==4.6.0
celery==5.3.1
flower==2.0.1  # Celery monitoring

# HTTP requests
requests==2.31.0

# Email
Flask-Mail==0.9.1

# Environment management
python-dotenv==1.0.0

# Validation
marshmallow==3.20.1
Flask-Marshmallow==0.15.0

# Utilities
bcrypt==4.0.1
PyJWT==2.8.0

# Development
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0

# Production
gunicorn==21.2.0

# External APIs
openai==0.28.0
google-generativeai==0.2.0
